Specify a minimum Test::TCP version
There are test failures due to a missing export of listen_socket in
Net::EmptyPort. Since that module has no version, we specify a Test::TCP
minimum version.

The listen_port function was added in 2.15, but later versions fix some
Windows issues, so we use the latest version in order to minimize
test failures.